% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arrayBlocks.R
\name{arrayBlocks}
\alias{arrayBlocks}
\title{Blocks information of tables with dimension >= 3.}
\usage{
arrayBlocks(table)
}
\arguments{
\item{table}{A sample of data with dimensions >= 3, can be a matrix, an array
or a table.}
}
\value{
a list of block names and dimension infomation to call each block.
}
\description{
Blocks information of tables with dimension >= 3.
}
\examples{
Titanic
arrayBlocks(Titanic)
Titanic[,,1,1] # the first row of $blockTablecom correponds to the first block
Titanic[,,1,2] # the third row of $blockTablecom correponds to the third block
}
\author{
R. Wayne Oldford and Xiaomei Yu
}

