% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chooseValue.R
\name{chooseValue}
\alias{chooseValue}
\title{Get a summary of a dimension/combination of dimensions.}
\usage{
chooseValue(table, chooseFun = "median", valueFun = "mad", comleft = NULL,
  com = NULL, column = 1)
}
\arguments{
\item{table}{A sample of data with dimensions >=2, can be a matrix, an array
or a table.}

\item{chooseFun}{The summary to represent the variability (spread) of each
dimension/combination of dimensions, can be functions in R or self-defined
functions.}

\item{valueFun}{The summary to measure the variability (spread) of each layer of the
dimension/combination of dimensions, can be functions in R or
self-defined functions.}

\item{comleft}{A matrix of combinations of dimensions, each column stores the
dimensions of a combination.}

\item{com}{A matrix of all the layer combinations of the dimensions in the combination
combination. Each column stores a combination of layers e.g. The dimensions in a
combination is (2,4,1), if the first column of com is (1,2,4), this means
the layer combination is the first layer of the second dimension and
the second layer of the fourth dimension and the fourth layer of the first
dimension.}

\item{column}{The column of comleft which we want to apply our chooseValue()
to, the default is 1.}
}
\value{
The summary of the dimension/combination of dimensions
}
\description{
Get a summary of a dimension/combination of dimensions.
}
\note{
Here, the default of chooseFun is "median" and the default of valueFun
  is "mad" since in our tidytable package we define the varaibility of a dimension
  (combination of dimensions) as the median of mad values get from each layer
  of that dimension (combination).
}
\examples{
table <- Titanic
comleft <- combn(length(dim(table)), 2) # all the combinations of two dimensions
comleft <- matrix(comleft, ncol = dim(comleft)[2])
# Get the summary of the second combination (second column of comleft), i.e. (1,3)
# use expand.grid() to get all the layer combinations
excom <- mapply(function(i){
paste('1:dim(table)[comleft[', i, ',2]]')
},
1:dim(comleft)[1]
)
com <- eval(parse(text = paste('expand.grid(',
paste(excom, collapse = ","),
')',
sep = ''
)
)
)
chooseValue(table, chooseFun = "median", valueFun="mad", comleft, com, column=2)
}
\author{
R. Wayne Oldford and Xiaomei Yu
}

