% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scientificNotation.R
\name{countUnique}
\alias{countUnique}
\title{Get the unique values and count the numbers of them in a sample}
\usage{
countUnique(x)
}
\arguments{
\item{x}{A sample of data}
}
\value{
A two-column matrix, the first column records the unique values
        and the second column are the correponding numbers of the unique values
}
\description{
Get the unique values and count the numbers of them in a sample
}
\examples{
countUnique(c(1,4,5,2,1,1,4,2))
}
\author{
Wayne Oldford and Xiaomei Yu
}

