% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findCombination.R
\name{findCombination}
\alias{findCombination}
\title{Find a dimension/combination of dimensions with certain conditions.}
\usage{
findCombination(table, combs = "NULL", chooseFun = "median",
  valueFun = "mad", dimDecreasing = TRUE)
}
\arguments{
\item{table}{A sample of data with dimensions >=2, can be a matrix, an array
or a table.}

\item{combs}{A matrix of dimension combinations, each column stores the
dimensions of a combination.}

\item{chooseFun}{The summary to represent the variability (spread) of each
dimension/combination of dimensions, can be functions in R or self-defined
functions.}

\item{valueFun}{The summary to measure the variability (spread) of each layer
of the dimension/combination of dimensions, can be functions in R or
self-defined functions.}

\item{dimDecreasing}{Logical, if order the dimensions/combinations in a
decreasing order.}
}
\value{
The dimension/combination of dimensions which is ordered first.
}
\description{
Find a dimension/combination of dimensions with certain conditions.
}
\note{
Here, the default of chooseFun is "median" and the default of valueFun
  is "mad" since in our tidytable package we define the varaibility of a
  dimension (combination of dimensions) as the median of mad values get from
  each layer of that dimension (combination).
}
\examples{
table <- Titanic
coms <- combn(length(dim(table)), 2) # all the combinations of two dimensions
findCombination(table, coms, chooseFun = "median", valueFun = "mad", dimDecreasing = TRUE)
# so the combination of first two dimensions has the largest variability
# among all combinations of two dimensions in Titanic data
}
\author{
R. Wayne Oldford and Xiaomei Yu
}

