% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/showtidytable.R
\name{showtidytable}
\alias{showtidytable}
\title{Show (Print) tidy table in console or LaTex.}
\usage{
showtidytable(table, display = c("console", "latex"), tableCaption = NULL,
  file = NULL, append = FALSE, removeLocation = TRUE,
  printLocation = TRUE, locationType = c("common", "withinEachTwoWay"),
  locTrimFraction = 0, fixedOrderDims = NULL, orderSummary = "mean",
  orderDecreasing = TRUE, orderType = c("common", "withinEachTwoWay"),
  swapDimensions = TRUE, preserveFirst2dim = NULL, swapDecreasing = TRUE,
  chooseFun = "median", valueFun = "mad", swapType = c("common",
  "withinEachTwoWay"), removeCommonZeros = TRUE, printUnit = TRUE,
  unitType = c("common", "withinEachTwoWay"), nSig = 2, expsRange = nSig -
  2)
}
\arguments{
\item{table}{A sample of data with dimensions >= 2, can be a matrix, an array
or a table.}

\item{display}{To show the table in console or laTex.}

\item{tableCaption}{The caption of the table.}

\item{file}{If NULL, then the latex code will be shown in console. Otherwise,
the code will be written to file (as in the latex command from the Hmisc
package).}

\item{append}{Logical, append latex output to an existing file.}

\item{removeLocation}{Logical, detect and remove the location.}

\item{printLocation}{Logical, print the location.}

\item{locationType}{Whether to remove and print the location within each
two-way block.}

\item{locTrimFraction}{The fraction to be trimmed from each end of a sample
to find the location.}

\item{fixedOrderDims}{A vector of those dimensions within which the layers
cannot be reordered.}

\item{orderSummary}{The summary which we use to order the dimension layerss.}

\item{orderDecreasing}{Logical, if the order of dimension elements is
decreasing.}

\item{orderType}{Whether to order the rows and columns within each two-way
block.}

\item{swapDimensions}{Logical, reorder dimensions (swap).}

\item{preserveFirst2dim}{Fix the first 2 dimensions of the table to be a
specific two dimensions of the table.}

\item{swapDecreasing}{Logical, if swap the dimensions in a way of spread
decreasing.}

\item{chooseFun}{The summary to represent the variability (spread) of each
dimension/combination of dimensions, can be functions in R or self-defined
functions.}

\item{valueFun}{The summary to measure the variability (spread) of each layer
of the dimension/combination of dimensions, can be functions in R or
self-defined functions.}

\item{swapType}{Whether to swap the rows and columns within each two-way
block.}

\item{removeCommonZeros}{Logical, remove common rightmost zeros from all
table entries.(Find a unit)}

\item{printUnit}{Logical, whether to print the location.}

\item{unitType}{Whether to remove common zeros and print unit within each
two-way block.}

\item{nSig}{Number of digits showed in the table.}

\item{expsRange}{The threshold to allow how many numbers to be showed. If the
spread of exponents of the table is smaller than expsRange, then all the
numbers in the table will be showed in the new table. Otherwise, only those
numbers whose exponents are within this expsRange will be displayed and we
will try to display most of the numbers.}
}
\value{
The tidy table in console or laTex format.
}
\description{
Show (Print) tidy table in console or LaTex.
}
\examples{
table <- array(c(5*10^5,9*10^5,2*10^5,8*10^5,
5*10^4,9*10^4,2*10^4,8*10^4,
5*10^2,9*10^2,2*10^2,8*10^2,
5*10^3,9*10^3,2*10^3,8*10^3,
7*10^9,4*10^9,2*10^9,5*10^9,
8*10^2,3*10^2,7*10^2,1*10^2,
6*10^3,8*10^3,2*10^3,9*10^3,
4*10^7,9*10^7,3*10^7,1*10^7
), c(2,2,2,2,2))
dimnames(table) <- list(c("i1","i2"),c("j1","j2"),c("k1","k2"),c("l1","l2"),c("m1","m2"))
table
tidytable(table)
showtidytable(table,display="console",tableCaption="tidy table",
locationType="withinEachTwoWay", orderType="withinEachTwoWay",
swapType="withinEachTwoWay", unitType="withinEachTwoWay")

showtidytable(table,display="latex",tableCaption="tidy table",
locationType="withinEachTwoWay", orderType="withinEachTwoWay",
swapType="withinEachTwoWay", unitType="withinEachTwoWay")


Titanic+264200
showtidytable(Titanic+264200,display="console",
tableCaption="Titanic tidy table", locationType="withinEachTwoWay",
orderType="withinEachTwoWay",swapType="withinEachTwoWay", unitType="withinEachTwoWay")

showtidytable(Titanic+264200,display="latex",tableCaption="Titanic tidy table",
 locationType="withinEachTwoWay", orderType="withinEachTwoWay",
 swapType="withinEachTwoWay", unitType="withinEachTwoWay")
}
\author{
R. Wayne Oldford and Xiaomei Yu
}

