% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tidytable.R
\name{tidytable}
\alias{tidytable}
\title{Generate a tidy table.}
\usage{
tidytable(table, removeLocation = TRUE, locTrimFraction = 0,
  fixedOrderDims = NULL, orderSummary = "mean", orderDecreasing = TRUE,
  swapDimensions = TRUE, preserveFirst2dim = NULL, swapDecreasing = TRUE,
  chooseFun = "median", valueFun = "mad", removeCommonZeros = TRUE,
  nSig = 2, expsRange = nSig - 1)
}
\arguments{
\item{table}{A sample of data with dimensions >=2, can be a matrix, an array
or a table.}

\item{removeLocation}{Logical, detect and remove the location.}

\item{locTrimFraction}{The fraction to be trimmed from each end of a sample
to find the location.}

\item{fixedOrderDims}{A vector of dimensions within which elements
cannot be reordered.}

\item{orderSummary}{The summary which we use to order the dimension layers.}

\item{orderDecreasing}{Logical, if the order of dimension elements is
decreasing.}

\item{swapDimensions}{Logical, reorder dimensions (swap).}

\item{preserveFirst2dim}{Fix the first 2 dimensions of the table to be a
specific two dimensions of the table.}

\item{swapDecreasing}{Logical, if swap the dimensions in a way of spread
decreasing.}

\item{chooseFun}{The summary to represent the variability (spread) of each
dimension/combination of dimensions, can be functions in R or self-defined
functions.}

\item{valueFun}{The summary to measure the variability (spread) of each layer
of the dimension/combination of dimensions, can be functions in R or
self-defined functions.}

\item{removeCommonZeros}{Logical, remove common rightmost zeros from all
table entries. (Find a unit.)}

\item{nSig}{Number of digits showed in the table.}

\item{expsRange}{The range of exponents within which
we hope to show most of the values,  default is nSig-1}
}
\value{
A list containing the named components:
        tidytable The tidy table,
        originalTable The original table,
        location The location removed from the original table,
        nSig Number of digits shown in the tidy table,
        units The units of the entries in the tidy table,
        newDimOrder The new dimension order after swapping the table.
}
\description{
Generate a tidy table.
}
\examples{
table1<-c(97.62,92.24,100.90,90.39,48.29,42.31,49.98,39.09,75.23,75.16,100.11,74.23,49.69,57.21,80.19,51.09)
table1 <- matrix(table1, nrow=4,byrow=TRUE,dimnames=list(c("North" ,"South","East","West"),c("Q1","Q2","Q3","Q4")))
table1
tidytable(table1)

table2 <- c(.7999, .7998, .7998, .7997,
3.7999,3.7824,3.7662,3.7223,
0.3,1.2,4.9,145.7,
20.799,20.699,20.899,145.699,
35.3, 34.5,33.6,34.7)
table2 <- matrix(table2, ncol=4,byrow=TRUE,dimnames=list(c("A" ,"B","C","D","E"),c("1","2","3","4")))
tidytable(table2)


UCBAdmissions
tidytable(UCBAdmissions)

HairEyeColor
tidytable(HairEyeColor)


Titanic
tidytable(Titanic)
}
\author{
R. Wayne Oldford and Xiaomei Yu
}

